Perform first successful gql query for getUser
Spent alot of time fixing errors in the
schema that caused parsing errors and
inability to connect to gql server while
using the playground. Learned that
we can't put an empty [] value inside
type definitions, and when we define
a query or mutation, it has to return
a value after the : semicolon.